# faster path intersects