… examples/gradient.png
… examples/heart.png
… examples/masking.png
… examples/shadow.png
… examples/text.png
… examples/text_spans.png
… examples/tiger.png
… tests/contexts/bezierCurveTo_1.png
… tests/contexts/clip_1.png
… tests/contexts/clip_1b.png
… tests/contexts/clip_1d.png
… tests/contexts/clip_1f.png
… tests/contexts/clip_3.png
… tests/contexts/ellipse_1.png
… tests/contexts/quadracticCurveTo_1.png
… tests/contexts/quadracticCurveTo_2.png
… tests/contexts/strokeText_1.png
… tests/fileformats/svg/diffs/Ghostscript_Tiger.png
… tests/fileformats/svg/diffs/circle01.png
… tests/fileformats/svg/diffs/ellipse01.png
… tests/fileformats/svg/diffs/miterlimit.png
… tests/fileformats/svg/diffs/polygon01.png
… tests/fileformats/svg/diffs/quad01.png
… tests/fileformats/svg/diffs/rect02.png
… tests/fileformats/svg/diffs/scale.png
… tests/fileformats/svg/emojitwo.png
… tests/fileformats/svg/flat-color-icons.png
… tests/fileformats/svg/ionicons.png
… tests/fileformats/svg/noto-emoji.png
… tests/fileformats/svg/openmoji.png