before:
`svg decode ........................ 18.642 ms     18.776 ms    ±0.098   x266`

after:
`svg decode ........................ 18.183 ms     18.360 ms    ±0.167   x272`